lab/gcp/16-01-2025-santa/requirements.txt
Flask==3.1.2
Flask-RESTful==0.3.10
google-cloud-firestore==2.22.0
WTForms==3.2.1
gunicorn==23.0.0

flask-cors==6.0.2